requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
anthropic>=0.40.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
